requirements.txt
sentence-transformers
tqdm
openpyxl
pandas
pyinstaller
